app/test_reverse_string_exclude_special.py
import pytest
from unittest.mock import patch
from string_reversal import reverse_string_exclude_special

def test_reverse_string_exclude_special_typical():
    # This test checks the function with a typical input string.
    result = reverse_string_exclude_special('abc!@#def')
    assert result == 'fed!@#cba', 'Expected output for typical input string'

def test_reverse_string_exclude_special_edge_case_empty():
    # This test checks the function with an empty string.
    result = reverse_string_exclude_special('')
    assert result == '', 'Expected output for empty string'

def test_reverse_string_exclude_special_edge_case_no_alnum():
    # This test checks the function with a string containing no alphanumeric characters.
    result = reverse_string_exclude_special('!!!@@@')
    assert result == '!!!@@@', 'Expected output for string with no alphanumeric characters'

def test_reverse_string_exclude_special_invalid_input():
    # This test checks the function with a non-string input.
    with pytest.raises(TypeError):
        reverse_string_exclude_special(12345)

def test_reverse_string_exclude_special_with_mock():
    @patch('string_reversal.some_external_dependency')
    def test_function_with_dependency(mock_dependency):
        mock_dependency.return_value = 'mocked_value'
        result = reverse_string_exclude_special('abc!@#def')
        assert result == 'fed!@#cba', 'Expected output when dependency is mocked'