requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.2
filetype==1.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.2.6
libmagic==1.0
lxml==5.3.0
marshmallow==3.23.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.58.1
orjson==3.10.12
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
psutil==6.1.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pypdf==5.1.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-magic-bin==0.4.14
python-oxmsg==0.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.11
unstructured-client==0.28.1
urllib3==2.3.0
webencodings==0.5.1
wrapt==1.17.0
yarl==1.18.3
